fix: Soft #712 max-lines — extract remote remove scripts
Move remove_worktree_script into remote_managed_workspace_remove_script
so remote_managed_workspace.rs stays under the 500-line ratchet after the
pullfrog live-branch guard.
